Renamign visitors -> tensor_collection
