+++ src/nn/conv.rs
@@ -3,7 +3,7 @@ use rand_distr::uniform::SampleUniform;
 
 use crate::{gradients::Tape, shapes::*, tensor::*, tensor_ops::*};
 
-use super::{visitors::*, BuildModule, BuildOnDevice, Module, ModuleMut, ToDevice};
+use super::{tensor_collection::*, BuildModule, BuildOnDevice, Module, ModuleMut, ToDevice};
 
 pub mod builder {
     #[derive(Debug)]
