+++ src/nn/repeated.rs
@@ -1,6 +1,6 @@
 use crate::{shapes::Dtype, tensor::*};
 
-use super::{visitors::*, BuildModule, BuildOnDevice, Module, ModuleMut, ToDevice};
+use super::{tensor_collection::*, BuildModule, BuildOnDevice, Module, ModuleMut, ToDevice};
 
 /// Repeats `T` `N` times. This requires that `T`'s input is the same as it's output.
 ///
