+++ src/nn/transformer/mod.rs
@@ -11,7 +11,7 @@ use rand_distr::uniform::SampleUniform;
 
 use crate::{shapes::*, tensor::*, tensor_ops::*};
 
-use super::{visitors::*, BuildModule, BuildOnDevice, Module, ModuleMut, ToDevice};
+use super::{tensor_collection::*, BuildModule, BuildOnDevice, Module, ModuleMut, ToDevice};
 
 pub mod builder {
     #[derive(Debug, Clone)]
